Replace babel with typescript
With typescript we no longer need the babel-loader to transpile ES-style
modules on the server, or to parse JSX. So I've cut out the babel
configurations in favor of ts-alternatives wherever possible.

Also swaps the babel-parser in eslint for the typescript parser, and
adds the typescript-eslint recommended rules, then tweaks a few things
that conflict.