Use webpack root element plugin
Adds an additional webpack plugin that hooks into the existing
html-webpack-plugin to add <div id="root" /> to the body of the
generated index.html so we no longer need to specify the path to the
template file.

In the process of adding this I found a weird typescript quirk where
defining the configuration in a variable beforehand resulted in a type
error on the "mode" value. Upshot is that the passing the configuration
directly into the webpack() function resolves it.