Switch to useReducer for managing messages
The original implementation with useState was a bit clunky, so I changed
it to useReducer for a more redux-like setup that made it easier to
reuse logic from the EAF app. I may revisit the file structure at some
point -- 'context' isn't totally accurate to describe everything in the
directory.