Better handling of tsconfig paths
Was missing the tsconfig path aliases in the dev.middleware. While
updating realized that there's a convenient webpack module to do the
same thing _without_ needing to maintain three lists of aliases, so
that's cool.

Also adds tsconfig-paths/register to the nodemon config so that paths
will be correctly resolved in server-side code as well. That should
allow for resolving paths everywhere in the codebase, he typed into the
commit message confidently.