Optimize webpack builds
The default "production" mode for webpack seems to be causing memory
exhaustion issues in travis (and occasionally on my local machine). I've
changed it to "none" for our production build and added back some of the
included configs/plugins to try to reduce the overhead and prevent
failures.

Godspeed, Travis.