Add token for docs; tag build
Forgot that we needed a github personal access token to allow us to push
generated documentation to the gh-pages branch. The actual token is
defined in the travis UI. I used one associated with my github account,
but we may want to create a service account for this in the future.

Additional, I added a better tag for the docker build based on
environment variables in travis.