Don't bundle assets on server
Current configuration wraps server dependencies into a single file with
the code. This ends up breaking a current dependency (xml-encryption)
because it uses fs.readFile to read in a template file, which does not
get picked up by the webpack tree. Rather than individually try to
catch every possible edge case like this, it's better to just rely on
the installed node_modules.