Separate typechecking from testing
Type errors emitted by ts-node will prevent tests from running but won't
be reported to the user. These errors can be ingored with the
TS_NODE_TRANSPILE_ONLY flag, but those are still things that we want to
detect. So I've added a separate typecheck script, which will be run on travis
before any linting or testing is done.