Replace `id` with `eppn`
Presence  of an `id` field on `User` has the potential to be misleading and suggest that it represents a database PK id. Since users are not stored in the database this is not the case. Instead, `eppn` is a much better name for the field and more accurately describes it's purpose