Basic webpack setup
Create basic client and server webpack config files along with top level
webpack config file to import and run both of them.

Both the client and the server have their own independently runnable
webpack config file, which contains config specific for that context.
This can then be imported into the top level config (and overridden if
required) as well as into a webpack dev server config file so that the
config to compile the app is the same as the config that runs it. This
results in a much cleaner webpack setup, as well as allowing more
complex things to be re-used and/or passed down like environment
variables.